Add files via upload
8-1-2020 KenUnix

BwbasicXP is for older 32 bit Windows/DOS systems. Like Windows XP and Windows 7 and Server 2000.
Bwbasic64 if for 64 bit Windows 10 systems like Windows 10.
bwbasic_3.20b-1j_i386.deb is for Intel 386 Linux systems.
bwbasic_3.20b-1j_armhf.deb is for 32 bit ARM Linux systems like Beglebone and C.H.I.P.
bwbasic_3.20b-1j_amd64.deb is for 64 bit AMD based Linux systems like Debian 10.4 and Ubuntu 20.04.

For the Windows packages download the executable and run it.
For the deb packages download the deb file and use dpkg -i <package name> to install it.

For the Windows packages Example programs are in \Program Files (X86)\Bwbasic\Examples
and the DOCS are in \Program Files (X86)\Bwbasic\DOCS.

For the deb installs the Examples are in  /usr/share/bwbasic
and the DOCS are in /usr/share/doc/bwbasic.